Added Compatibility Checker to reduce redundancy
